requirements.txt
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
pymysql
